+++ Assets/Scripts/Tower/Bullet.cs
@@ -7,7 +7,7 @@ public class Bullet
     [SerializeField]
     private float _speed;
     [SerializeField]
-    private float _damage;
+    private int _damage;
     [SerializeField]
     private string _targetTag;
     [SerializeField]
@@ -16,7 +16,7 @@ public class Bullet
     private float _critical_Damage;
 
     public float Speed { get => _speed; set => _speed = value; }
-    public float Damage { get => _damage; set => _damage = value; }
+    public int Damage { get => _damage; set => _damage = value; }
     public string TargetTag { get => _targetTag; set => _targetTag = value; }
     public float KnockbackDistance { get => _knockbackDistance; set => _knockbackDistance = value; }
     public float CriticalDamage { get => _critical_Damage; set => _critical_Damage = value; }
